+++ KaiFighter/KaiFighterGame/Objects/DynamicObjects/Characters/Enemies/Archer.cs
@@ -7,7 +7,7 @@
     {
         private const string CollisionGroupString = "Archer";
 
-        public Archer(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, int damage, int health, int cooldown)
+        public Archer(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, double damage, double health, int cooldown)
             : base(position, imageLocation, objectType, objColor, scale, rotation, layerDepth, movementSpeed, damage, health, cooldown)
         {
         }
