+++ KaiFighter/KaiFighterGame/Objects/DynamicObjects/Projectiles/Bullet.cs
@@ -10,14 +10,14 @@
         private Vector2 targetDir;
         private bool isPlayerFire;
 
-        public Bullet(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, int damage, Vector2 targetDirection)
+        public Bullet(Vector2 position, string imageLocation, ObjectType objectType, Color? objColor, float scale, float rotation, float layerDepth, float movementSpeed, double damage, Vector2 targetDirection)
             : base(position, imageLocation, objectType, objColor, scale, rotation, layerDepth, movementSpeed)
         {
             this.Damage = damage;
             this.targetDir = targetDirection;
         }
 
-        public int Damage { get; set; }
+        public double Damage { get; set; }
 
         // TODO - Change that !!!
         public bool IsPlayerFire
